% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Animals}
\alias{Animals}
\title{Animals}
\format{
A data frame with 28 rows and 2 columns:
\describe{
    \item{body}{numeric: ... }
    \item{brain}{numeric: ... }
}
}
\usage{
Animals
}
\description{
Animals
}
\keyword{datasets}
