% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{anorexia}
\alias{anorexia}
\title{anorexia}
\format{
A data frame with 72 rows and 3 columns:
\describe{
    \item{Treat}{factor: levels Cont (Control), CBT (Cognative Behaviour Therapy) and FT (Family Threapy) }
    \item{Prewt}{numeric: Weight before treatment (lbs) }
    \item{Postwt}{numeric: Weight after treatment (lbs) }
}
}
\usage{
anorexia
}
\description{
Data from a clinical trial for Anorexia patients.
}
\keyword{datasets}
