% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Aus}
\alias{Aus}
\title{Aus}
\format{
A data frame with 439470 rows and 3 columns:
\describe{
    \item{x}{numeric: Longitude }
    \item{y}{numeric: Latitude }
    \item{group}{character: A grouping factor for discrete segments}
}
}
\usage{
Aus
}
\description{
Australian coastline in a form suitable for ggplot graphics
}
\examples{
library(ggplot2)
tas <- subset(Aus, y < -39.5)
ggplot(tas) + aes(x = x, y = y, group = group) + geom_path(colour = "firebrick") + coord_equal() +
  labs(x = "Longitude", y = "Latitude", title = "Tasmania") + theme_bw() +
  theme(plot.title = element_text(hjust = 0.5))
rm(tas)
}
\keyword{datasets}
