% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{BeachBirds}
\alias{BeachBirds}
\title{BeachBirds}
\format{
A data frame with 483 rows and 5 columns:
\describe{
    \item{Site}{integer: ... }
    \item{Species}{factor: ... }
    \item{Sex}{factor: ... }
    \item{flush.dist}{numeric: ... }
    \item{land.dist}{numeric: ... }
}
}
\usage{
BeachBirds
}
\description{
BeachBirds
}
\keyword{datasets}
