% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{biopsy}
\alias{biopsy}
\title{biopsy}
\format{
A data frame with 699 rows and 11 columns:
\describe{
    \item{ID}{character: ... }
    \item{V1}{integer: ... }
    \item{V2}{integer: ... }
    \item{V3}{integer: ... }
    \item{V4}{integer: ... }
    \item{V5}{integer: ... }
    \item{V6}{integer: ... }
    \item{V7}{integer: ... }
    \item{V8}{integer: ... }
    \item{V9}{integer: ... }
    \item{class}{factor: ... }
}
}
\usage{
biopsy
}
\description{
biopsy
}
\keyword{datasets}
