% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{birthwt}
\alias{birthwt}
\title{birthwt}
\format{
A data frame with 189 rows and 10 columns:
\describe{
    \item{low}{integer: ... }
    \item{age}{integer: ... }
    \item{lwt}{integer: ... }
    \item{race}{integer: ... }
    \item{smoke}{integer: ... }
    \item{ptl}{integer: ... }
    \item{ht}{integer: ... }
    \item{ui}{integer: ... }
    \item{ftv}{integer: ... }
    \item{bwt}{integer: ... }
}
}
\usage{
birthwt
}
\description{
birthwt
}
\keyword{datasets}
