% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Boston}
\alias{Boston}
\title{Boston}
\format{
A data frame with 506 rows and 14 columns:
\describe{
    \item{crim}{numeric: ... }
    \item{zn}{numeric: ... }
    \item{indus}{numeric: ... }
    \item{chas}{integer: ... }
    \item{nox}{numeric: ... }
    \item{rm}{numeric: ... }
    \item{age}{numeric: ... }
    \item{dis}{numeric: ... }
    \item{rad}{integer: ... }
    \item{tax}{numeric: ... }
    \item{ptratio}{numeric: ... }
    \item{black}{numeric: ... }
    \item{lstat}{numeric: ... }
    \item{medv}{numeric: ... }
}
}
\usage{
Boston
}
\description{
Boston
}
\keyword{datasets}
