% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Cars93}
\alias{Cars93}
\title{Cars93}
\format{
A data frame with 93 rows and 27 columns:
\describe{
    \item{Manufacturer}{factor: ... }
    \item{Model}{factor: ... }
    \item{Type}{factor: ... }
    \item{Min.Price}{numeric: ... }
    \item{Price}{numeric: ... }
    \item{Max.Price}{numeric: ... }
    \item{MPG.city}{integer: ... }
    \item{MPG.highway}{integer: ... }
    \item{AirBags}{factor: ... }
    \item{DriveTrain}{factor: ... }
    \item{Cylinders}{factor: ... }
    \item{EngineSize}{numeric: ... }
    \item{Horsepower}{integer: ... }
    \item{RPM}{integer: ... }
    \item{Rev.per.mile}{integer: ... }
    \item{Man.trans.avail}{factor: ... }
    \item{Fuel.tank.capacity}{numeric: ... }
    \item{Passengers}{integer: ... }
    \item{Length}{integer: ... }
    \item{Wheelbase}{integer: ... }
    \item{Width}{integer: ... }
    \item{Turn.circle}{integer: ... }
    \item{Rear.seat.room}{numeric: ... }
    \item{Luggage.room}{integer: ... }
    \item{Weight}{integer: ... }
    \item{Origin}{factor: ... }
    \item{Make}{factor: ... }
}
}
\usage{
Cars93
}
\description{
Cars93
}
\keyword{datasets}
