% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{cats}
\alias{cats}
\title{cats}
\format{
A data frame with 144 rows and 3 columns:
\describe{
    \item{Sex}{factor: ... }
    \item{Bwt}{numeric: ... }
    \item{Hwt}{numeric: ... }
}
}
\usage{
cats
}
\description{
cats
}
\keyword{datasets}
