% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{crabs}
\alias{crabs}
\title{crabs}
\format{
A data frame with 200 rows and 8 columns:
\describe{
    \item{sp}{factor: ... }
    \item{sex}{factor: ... }
    \item{index}{integer: ... }
    \item{FL}{numeric: ... }
    \item{RW}{numeric: ... }
    \item{CL}{numeric: ... }
    \item{CW}{numeric: ... }
    \item{BD}{numeric: ... }
}
}
\usage{
crabs
}
\description{
crabs
}
\keyword{datasets}
