% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{creditCards}
\alias{creditCards}
\title{creditCards}
\format{
A data frame with 2085 rows and 65 columns:
\describe{
    \item{address.changes}{integer: (self describing)}
    \item{address.language}{factor: (self describing)}
    \item{address.lang.changes}{integer: (self describing)}
    \item{current.profession}{factor: (self describing)}
    \item{profession.changes}{integer: (self describing)}
    \item{gender}{factor: (self describing)}
    \item{num.gender.corrections}{integer: (self describing)}
    \item{name.changes}{integer: (self describing)}
    \item{current.nationality}{factor: (self describing)}
    \item{nationality.changes}{integer: (self describing)}
    \item{phone.changes}{integer: (self describing)}
    \item{cust.age}{numeric: (self describing)}
    \item{credit.card.owner}{factor: (self describing)}
    \item{mean.num.atm.withdr}{numeric: (self describing)}
    \item{mean.num.check.cash.withdr}{numeric: (self describing)}
    \item{mean.num.check.cash.deposits}{numeric: (self describing)}
    \item{mean.num.reg.pmnt.init.by.cust}{numeric: (self describing)}
    \item{mean.num.salary.deposits}{numeric: (self describing)}
    \item{mean.num.transfers}{numeric: (self describing)}
    \item{mean.amnt.pmnts.init.by.cust}{numeric: (self describing)}
    \item{mean.num.security.pur.ord}{numeric: (self describing)}
    \item{mean.num.security.sales.ord}{numeric: (self describing)}
    \item{mean.amnt.atm.withdr}{numeric: (self describing)}
    \item{mean.check.cash.withdr}{numeric: (self describing)}
    \item{mean.cash.deposits}{numeric: (self describing)}
    \item{mean.amnt.reg.pmnt.init.by.cust}{numeric: (self describing)}
    \item{mean.check.credits}{numeric: (self describing)}
    \item{mean.salary.deposits}{numeric: (self describing)}
    \item{mean.check.debits}{numeric: (self describing)}
    \item{mean.amnt.transfers}{numeric: (self describing)}
    \item{mean.num.pmnts.init.by.cust}{numeric: (self describing)}
    \item{mean.amnt.security.pur.ord}{numeric: (self describing)}
    \item{mean.amnt.security.sales.ord}{numeric: (self describing)}
    \item{mean.num.saving.cash.withdr}{numeric: (self describing)}
    \item{mean.num.saving.cash.deposits}{numeric: (self describing)}
    \item{mean.amnt.saving.cash.withdr}{numeric: (self describing)}
    \item{mean.amnt.saving.cash.deposits}{numeric: (self describing)}
    \item{mean.saving.balance}{numeric: (self describing)}
    \item{stdev.num.atm.withdr}{numeric: (self describing)}
    \item{stdev.num.check.cash.withdr}{numeric: (self describing)}
    \item{stdev.num.check.cash.deposits}{numeric: (self describing)}
    \item{stdev.num.reg.pmnt.init.by.cust}{numeric: (self describing)}
    \item{stdev.num.salary.deposits}{numeric: (self describing)}
    \item{stdev.num.transfers}{numeric: (self describing)}
    \item{stdev.amnt.pmnts.init.by.cust}{numeric: (self describing)}
    \item{stdev.num.security.pur.ord}{numeric: (self describing)}
    \item{stdev.num.security.sales.ord}{numeric: (self describing)}
    \item{stdev.amnt.atm.withdr}{numeric: (self describing)}
    \item{stdev.check.cash.withdr}{numeric: (self describing)}
    \item{stdev.cash.deposits}{numeric: (self describing)}
    \item{stdev.amnt.reg.pmnt.init.by.cust}{numeric: (self describing)}
    \item{stdev.check.credits}{numeric: (self describing)}
    \item{stdev.salary.deposits}{numeric: (self describing)}
    \item{stdev.check.debits}{numeric: (self describing)}
    \item{stdev.amnt.transfers}{numeric: (self describing)}
    \item{stdev.num.pmnts.init.by.cust}{numeric: (self describing)}
    \item{stdev.amnt.security.pur.ord}{numeric: (self describing)}
    \item{stdev.amnt.security.sales.ord}{numeric: (self describing)}
    \item{stdev.num.saving.cash.withdr}{numeric: (self describing)}
    \item{stdev.num.saving.cash.deposits}{numeric: (self describing)}
    \item{stdev.amnt.saving.cash.withdr}{numeric: (self describing)}
    \item{stdev.amnt.saving.cash.deposits}{numeric: (self describing)}
    \item{stdev.saving.balance}{numeric: (self describing)}
    \item{cust.age.bin}{factor: (self describing)}
    \item{mean.salary.deposits.bin}{factor: (self describing)}
}
}
\usage{
creditCards
}
\description{
creditCards
}
\keyword{datasets}
