% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{DarwinsFinches}
\alias{DarwinsFinches}
\title{DarwinsFinches}
\format{
A data frame with 146 rows and 10 columns:
\describe{
    \item{Id}{character: within species bird identifier}
    \item{Species}{factor: Species }
    \item{BodyL}{integer: Body length }
    \item{WingL}{numeric: Wing length }
    \item{TailL}{numeric: Tail length }
    \item{BeakW}{numeric: Beak width }
    \item{BeakH}{numeric: Beal height }
    \item{LBeakL}{numeric: Lower beak length }
    \item{UBeakL}{numeric: Upper beak length }
    \item{N.UBkL}{numeric: (not sure!) }
    \item{TarsusL}{numeric: Tarsus length }
}
}
\usage{
DarwinsFinches
}
\description{
Source: Snodgrass R and Heller E (1904) Papers from the Hopkins-Stanford Galapagos Expedition, 1898-99.
XVI. Birds. Proceedings of the Washington Academy of Sciences 5: 231-372.
}
\details{
All measures are in mm.
}
\keyword{datasets}
