% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{dutchSpeakers}
\alias{dutchSpeakers}
\title{dutchSpeakers}
\format{
A data frame with 80 rows and 6 columns:
\describe{
    \item{Ident}{character: ... }
    \item{Sex}{factor: ... }
    \item{AgeGroup}{factor: ... }
    \item{Region}{factor: ... }
    \item{Drawl}{numeric: ... }
    \item{Prolix}{numeric: ... }
}
}
\usage{
dutchSpeakers
}
\description{
dutchSpeakers
}
\keyword{datasets}
