% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{genotype}
\alias{genotype}
\title{genotype}
\format{
A data frame with 61 rows and 3 columns:
\describe{
    \item{Litter}{factor: ... }
    \item{Mother}{factor: ... }
    \item{Wt}{numeric: ... }
}
}
\usage{
genotype
}
\description{
genotype
}
\keyword{datasets}
