% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{GermanLoan}
\alias{GermanLoan}
\title{GermanLoan}
\format{
A data frame with 1000 rows and 21 columns:
\describe{
    \item{Cheque}{factor: ... }
    \item{Duration}{integer: ... }
    \item{History}{factor: ... }
    \item{Purpose}{factor: ... }
    \item{Amount}{integer: ... }
    \item{Savings}{factor: ... }
    \item{Employment}{factor: ... }
    \item{Installment}{integer: ... }
    \item{Personal}{factor: ... }
    \item{Other}{factor: ... }
    \item{Residence}{integer: ... }
    \item{Property}{factor: ... }
    \item{Age}{integer: ... }
    \item{Plans}{factor: ... }
    \item{Housing}{factor: ... }
    \item{Cards}{integer: ... }
    \item{Job}{factor: ... }
    \item{Dependents}{integer: ... }
    \item{Phone}{factor: ... }
    \item{Foreigner}{factor: ... }
    \item{Loan}{factor: ... }
}
}
\usage{
GermanLoan
}
\description{
See vignette for details.  Data supplied from
\url{http://archive.ics.uci.edu/ml/machine-learning-databases/statlog/german/}
}
\keyword{datasets}
