% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{geyser}
\alias{geyser}
\title{geyser}
\format{
A data frame with 299 rows and 2 columns:
\describe{
    \item{waiting}{numeric: ... }
    \item{duration}{numeric: ... }
}
}
\usage{
geyser
}
\description{
geyser
}
\keyword{datasets}
