% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{hankinson}
\alias{hankinson}
\title{hankinson}
\format{
A data frame with 11 rows and 2 columns:
\describe{
    \item{thetaDeg}{numeric: ... }
    \item{fTheta}{integer: ... }
}
}
\usage{
hankinson
}
\description{
hankinson
}
\keyword{datasets}
