% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{housing}
\alias{housing}
\title{housing}
\format{
A data frame with 72 rows and 5 columns:
\describe{
    \item{Sat}{ordered: ... }
    \item{Infl}{factor: ... }
    \item{Type}{factor: ... }
    \item{Cont}{factor: ... }
    \item{Freq}{integer: ... }
}
}
\usage{
housing
}
\description{
housing
}
\keyword{datasets}
