% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{immer}
\alias{immer}
\title{immer}
\format{
A data frame with 30 rows and 4 columns:
\describe{
    \item{Loc}{factor: ... }
    \item{Var}{factor: ... }
    \item{Y1}{numeric: ... }
    \item{Y2}{numeric: ... }
}
}
\usage{
immer
}
\description{
immer
}
\keyword{datasets}
