% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Iowa}
\alias{Iowa}
\title{Iowa}
\format{
A data frame with 33 rows and 10 columns:
\describe{
    \item{Year}{integer: ... }
    \item{Rain0}{numeric: ... }
    \item{Temp1}{numeric: ... }
    \item{Rain1}{numeric: ... }
    \item{Temp2}{numeric: ... }
    \item{Rain2}{numeric: ... }
    \item{Temp3}{numeric: ... }
    \item{Rain3}{numeric: ... }
    \item{Temp4}{numeric: ... }
    \item{Yield}{numeric: ... }
}
}
\usage{
Iowa
}
\description{
Iowa
}
\keyword{datasets}
