% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{janka}
\alias{janka}
\title{janka}
\format{
A data frame with 36 rows and 2 columns:
\describe{
    \item{Density}{numeric: ... }
    \item{Hardness}{integer: ... }
}
}
\usage{
janka
}
\description{
janka
}
\keyword{datasets}
