% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Janka2012}
\alias{Janka2012}
\title{Janka2012}
\format{
A data frame with 59 rows and 10 columns:
\describe{
    \item{Name}{factor: ... }
    \item{Colour}{factor: ... }
    \item{Binomial}{factor: ... }
    \item{Location}{factor: ... }
    \item{Lyctus}{factor: ... }
    \item{Hardness}{numeric: ... }
    \item{Density}{integer: ... }
    \item{Class}{integer: ... }
    \item{H2}{integer: ... }
    \item{Eucalypt}{logical: ... }
}
}
\usage{
Janka2012
}
\description{
Janka2012
}
\keyword{datasets}
