% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{lcmsms}
\alias{lcmsms}
\title{lcmsms}
\format{
A data frame with 140 rows and 4 columns:
\describe{
    \item{Sample}{factor: ... }
    \item{FB1}{numeric: ... }
    \item{FB2}{numeric: ... }
    \item{FB3}{numeric: ... }
}
}
\usage{
lcmsms
}
\description{
lcmsms
}
\keyword{datasets}
