% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Lignin}
\alias{Lignin}
\title{Lignin}
\format{
A data frame with 60 rows and 3 columns:
\describe{
    \item{Soluble}{numeric: ... }
    \item{Alkali}{integer: ... }
    \item{Permanganate}{numeric: ... }
}
}
\usage{
Lignin
}
\description{
Lignin
}
\keyword{datasets}
