% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRoxyData.R
\name{makeRoxyData}
\alias{makeRoxyData}
\alias{print.RoxyData}
\title{Make Roxygen Template for Data}
\usage{
makeRoxyData(dat)

\method{print}{RoxyData}(x, ...)
}
\arguments{
\item{dat}{A character string giving the name of the data frame}

\item{x}{An object of class "RoxyData"}

\item{...}{Additional arguments (currently not used)}
}
\value{
A character string vector, of class "RoxyData", with the roxygen template. By
    default, the output is copied to the clipboard and may be pasted into an R
    script directly.
}
\description{
Generate Roxygen template for a data frame object
}
\examples{
makeRoxyData("janka")
}
