% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{mammals}
\format{
A data frame with 62 rows and 2 columns:
\describe{
    \item{body}{numeric: ... }
    \item{brain}{numeric: ... }
}
}
\usage{
mammals
}
\description{
mammals
}
\keyword{datasets}
