% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{mcycle}
\alias{mcycle}
\title{mcycle}
\format{
A data frame with 133 rows and 2 columns:
\describe{
    \item{times}{numeric: ... }
    \item{accel}{numeric: ... }
}
}
\usage{
mcycle
}
\description{
mcycle
}
\keyword{datasets}
