% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{menarche}
\alias{menarche}
\title{menarche}
\format{
A data frame with 25 rows and 3 columns:
\describe{
    \item{Age}{numeric: ... }
    \item{Total}{numeric: ... }
    \item{Menarche}{numeric: ... }
}
}
\usage{
menarche
}
\description{
menarche
}
\keyword{datasets}
