% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{michelson}
\alias{michelson}
\title{michelson}
\format{
A data frame with 100 rows and 3 columns:
\describe{
    \item{Speed}{integer: ... }
    \item{Run}{factor: ... }
    \item{Expt}{factor: ... }
}
}
\usage{
michelson
}
\description{
michelson
}
\keyword{datasets}
