% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{muscle}
\alias{muscle}
\title{muscle}
\format{
A data frame with 60 rows and 3 columns:
\describe{
    \item{Strip}{factor: ... }
    \item{Conc}{numeric: ... }
    \item{Length}{numeric: ... }
}
}
\usage{
muscle
}
\description{
muscle
}
\keyword{datasets}
