% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{NRL}
\alias{NRL}
\title{NRL}
\format{
A character string vector.
}
\usage{
NRL
}
\description{
A character string with the names of the Australian
Rugby League competition teams
}
\keyword{datasets}
