% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{oats}
\alias{oats}
\title{oats}
\format{
A data frame with 72 rows and 4 columns:
\describe{
    \item{B}{factor: ... }
    \item{V}{factor: ... }
    \item{N}{factor: ... }
    \item{Y}{integer: ... }
}
}
\usage{
oats
}
\description{
oats
}
\keyword{datasets}
