% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Olympic}
\alias{Olympic}
\title{Olympic}
\format{
A data frame with 86 rows and 9 columns:
\describe{
    \item{Rank}{integer: ... }
    \item{Nation}{factor: ... }
    \item{Code}{factor: ... }
    \item{Gold}{integer: ... }
    \item{Silver}{integer: ... }
    \item{Bronze}{integer: ... }
    \item{Total}{integer: ... }
    \item{GDP.Per.Capita.US.}{integer: ... }
    \item{Population..thousands.}{integer: ... }
}
}
\usage{
Olympic
}
\description{
Olympic
}
\keyword{datasets}
