% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Oz}
\alias{Oz}
\title{Australian and nearby coastline coordinates}
\format{
A list with two numerical components, each of length 439740:
\describe{
  \item{x}{Longitude in decimal degrees, to 4dp.}
  \item{y}{Latitude in decimal degrees, to 4dp.}
}
}
\usage{
Oz
}
\description{
The coastline is given as map coordinates with
missing values inserted to indicate breaks in the coastline.
}
\examples{
if("WWRGraphics" \%in\% loadedNamespaces()) {
plot(Latitude ~ Longitude, roundTrip, asp = 1, ann = FALSE, bty="n")
lines(Oz, col = alpha("#8B4513", 0.5))
text(Latitude ~ Longitude, roundTrip, Locality, col = "#006400", xpd=NA,
     pos = avoid(Longitude, Latitude), cex=0.7, offset = 0.25, font=2)
} else cat("Attach WWRGraphics package first\n")
}
\keyword{datasets}
