% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{painters}
\alias{painters}
\title{painters}
\format{
A data frame with 54 rows and 5 columns:
\describe{
    \item{Composition}{integer: ... }
    \item{Drawing}{integer: ... }
    \item{Colour}{integer: ... }
    \item{Expression}{integer: ... }
    \item{School}{factor: ... }
}
}
\usage{
painters
}
\description{
painters
}
\keyword{datasets}
