% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{petrol}
\alias{petrol}
\title{petrol}
\format{
A data frame with 32 rows and 6 columns:
\describe{
    \item{No}{factor: ... }
    \item{SG}{numeric: ... }
    \item{VP}{numeric: ... }
    \item{V10}{integer: ... }
    \item{EP}{integer: ... }
    \item{Y}{numeric: ... }
}
}
\usage{
petrol
}
\description{
petrol
}
\keyword{datasets}
