% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{quine}
\alias{quine}
\title{quine}
\format{
A data frame with 146 rows and 5 columns:
\describe{
    \item{Eth}{factor: ... }
    \item{Sex}{factor: ... }
    \item{Age}{factor: ... }
    \item{Lrn}{factor: ... }
    \item{Days}{integer: ... }
}
}
\usage{
quine
}
\description{
quine
}
\keyword{datasets}
