% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{roundTrip}
\alias{roundTrip}
\title{Some localities in Australia}
\format{
A data frame with 29 rows and 4 columns:
\describe{
  \item{Locality}{Location name}
  \item{Longitude,Latitude}{Coordinates, in decimal degrees}
  \item{Population}{Approximate population in 2016}
}
}
\usage{
roundTrip
}
\description{
A data frame containing names of a sequence of
localities around Australia, as well as their
coordinates and approximate populaton.
}
\examples{
if("WWRGraphics" \%in\% loadedNamespaces()) {
z <- with(roundTrip, complex(,Longitude,Latitude) \%>\% setNames(Locality))
plot(z, asp=1, ann=FALSE, axes=FALSE)
arrows(z, cyc(z), col = "steel blue", gap = 1)
dis <- gcd_km(z, cyc(z)) \%>\% round
text((z + cyc(z))/2, dis, col = "red", cex = 0.8, font = 4)
lines(Oz, col = alpha("dark green", 0.5))
circles(Latitude ~ Longitude, roundTrip, radii = sqrt(Population),
        fill = "#FFC0CB80", colour = "hotpink")
text(z, names(z), pos = avoid(z), cex = 0.7, xpd = NA)
} else cat("Attach WWRGraphics package first\n")
}
\keyword{datasets}
