% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{SECoastline}
\alias{SECoastline}
\title{SECoastline}
\format{
A data frame with 395650 rows and 2 columns:
\describe{
    \item{x}{numeric: ... }
    \item{y}{numeric: ... }
}
plot(SECoastline, type = "l", asp = 1) ## may take some time
}
\usage{
SECoastline
}
\description{
Australian coastline with islands to the north and east,
including Papua New Guinea, Indonesia and New Zealand
}
\keyword{datasets}
