% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Set1}
\alias{Set1}
\title{Extended colour names}
\format{
A data frame with 931 rows and 2 columns:
\describe{
  \item{Color}{Common or commercial name of a colour}
  \item{Hex}{hex string defining the colour itself}
}
}
\usage{
Set1
}
\description{
A data frame containing names and hex values for an
extended list of colours.
}
\examples{
if("WWRGraphics" \%in\% loadedNamespaces()) {
  french <- getColours("French") \%>\% showColours()
} else cat("Attach WWRGraphics package first\n")

}
\keyword{datasets}
