% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Set2}
\alias{Set2}
\title{Extended colour names}
\format{
A data frame with 1746 rows and 2 columns:
\describe{
  \item{Color}{Common or commercial name of a colour}
  \item{Hex}{hex string defining the colour itself}
}
}
\usage{
Set2
}
\description{
A data frame containing names and hex values for an
extended list of colours.
}
\examples{
if("WWRGraphics" \%in\% loadedNamespaces()) {
  french <- getColours("French", "Set2") \%>\% showColours()
} else cat("Attach WWRGraphics package first\n")
}
\keyword{datasets}
