% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{SpamBase}
\alias{SpamBase}
\title{SpamBase}
\format{
A data frame with 4601 rows and 58 columns:
\describe{
    \item{wordFreqMake}{numeric: (self describing)}
    \item{wordFreqAddress}{numeric: (self describing)}
    \item{wordFreqAll}{numeric: (self describing)}
    \item{wordFreq3d}{numeric: (self describing)}
    \item{wordFreqOur}{numeric: (self describing)}
    \item{wordFreqOver}{numeric: (self describing)}
    \item{wordFreqRemove}{numeric: (self describing)}
    \item{wordFreqInternet}{numeric: (self describing)}
    \item{wordFreqOrder}{numeric: (self describing)}
    \item{wordFreqMail}{numeric: (self describing)}
    \item{wordFreqReceive}{numeric: (self describing)}
    \item{wordFreqWill}{numeric: (self describing)}
    \item{wordFreqPeople}{numeric: (self describing)}
    \item{wordFreqReport}{numeric: (self describing)}
    \item{wordFreqAddresses}{numeric: (self describing)}
    \item{wordFreqFree}{numeric: (self describing)}
    \item{wordFreqBusiness}{numeric: (self describing)}
    \item{wordFreqEmail}{numeric: (self describing)}
    \item{wordFreqYou}{numeric: (self describing)}
    \item{wordFreqCredit}{numeric: (self describing)}
    \item{wordFreqYour}{numeric: (self describing)}
    \item{wordFreqFont}{numeric: (self describing)}
    \item{wordFreq000}{numeric: (self describing)}
    \item{wordFreqMoney}{numeric: (self describing)}
    \item{wordFreqHp}{numeric: (self describing)}
    \item{wordFreqHpl}{numeric: (self describing)}
    \item{wordFreqGeorge}{numeric: (self describing)}
    \item{wordFreq650}{numeric: (self describing)}
    \item{wordFreqLab}{numeric: (self describing)}
    \item{wordFreqLabs}{numeric: (self describing)}
    \item{wordFreqTelnet}{numeric: (self describing)}
    \item{wordFreq857}{numeric: (self describing)}
    \item{wordFreqData}{numeric: (self describing)}
    \item{wordFreq4}{numeric: (self describing)}
    \item{wordFreq85}{numeric: (self describing)}
    \item{wordFreqTechnology}{numeric: (self describing)}
    \item{wordFreq}{numeric: (self describing)}
    \item{wordFreqParts}{numeric: (self describing)}
    \item{wordFreqPm}{numeric: (self describing)}
    \item{wordFreqDirect}{numeric: (self describing)}
    \item{wordFreqCs}{numeric: (self describing)}
    \item{wordFreqMeeting}{numeric: (self describing)}
    \item{wordFreqOriginal}{numeric: (self describing)}
    \item{wordFreqProject}{numeric: (self describing)}
    \item{wordFreqRe}{numeric: (self describing)}
    \item{wordFreqEdu}{numeric: (self describing)}
    \item{wordFreqTable}{numeric: (self describing)}
    \item{wordFreqConference}{numeric: (self describing)}
    \item{charFreqSemi}{numeric: (self describing)}
    \item{charFreqParn}{numeric: (self describing)}
    \item{charFreqBrak}{numeric: (self describing)}
    \item{charFreqBang}{numeric: (self describing)}
    \item{charFreqDoll}{numeric: (self describing)}
    \item{charFreqHash}{numeric: (self describing)}
    \item{capitalRunLengthAverage}{numeric: (self describing)}
    \item{capitalRunLengthLongest}{integer: (self describing)}
    \item{capitalRunLengthTotal}{integer: (self describing)}
    \item{isSpam}{factor: (self describing)}
}
}
\usage{
SpamBase
}
\description{
A data set got by classifying a large number of email
massages as to whether they were spam or not, together
with a large number of textual features of the message.
}
\keyword{datasets}
