% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{stormer}
\alias{stormer}
\title{stormer}
\format{
A data frame with 23 rows and 3 columns:
\describe{
    \item{Viscosity}{numeric: ... }
    \item{Wt}{integer: ... }
    \item{Time}{numeric: ... }
}
}
\usage{
stormer
}
\description{
stormer
}
\keyword{datasets}
