% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{Stormer}
\alias{Stormer}
\title{Stormer}
\format{
A data frame with 23 rows and 3 columns:
\describe{
    \item{Viscosity}{numeric: ... }
    \item{Weight}{integer: ... }
    \item{Time}{numeric: ... }
}
}
\usage{
Stormer
}
\description{
Stormer
}
\keyword{datasets}
