% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{talkers}
\alias{talkers}
\title{talkers}
\format{
A data frame with 80 rows and 6 columns:
\describe{
    \item{id}{integer: ... }
    \item{sex}{integer: ... }
    \item{age}{integer: ... }
    \item{region}{factor: ... }
    \item{syldur}{numeric: ... }
    \item{nsyl}{numeric: ... }
}
}
\usage{
talkers
}
\description{
talkers
}
\keyword{datasets}
