% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{titanicPassengers}
\alias{titanicPassengers}
\title{titanicPassengers}
\format{
A data frame with 891 rows and 12 columns:
\describe{
    \item{PassengerId}{integer: ... }
    \item{Survived}{character: ... }
    \item{Pclass}{character: ... }
    \item{Name}{character: ... }
    \item{Sex}{character: ... }
    \item{Age}{integer: ... }
    \item{SibSp}{integer: ... }
    \item{Parch}{integer: ... }
    \item{Ticket}{character: ... }
    \item{Fare}{numeric: ... }
    \item{Cabin}{character: ... }
    \item{Embarked}{character: ... }
}
}
\usage{
titanicPassengers
}
\description{
titanicPassengers
}
\keyword{datasets}
