% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{vinho_verde}
\alias{vinho_verde}
\title{vinho_verde}
\format{
A data frame with 4898 rows and 12 columns:
\describe{
     \item{fixed_acidity}{A numeric variable, ...}
     \item{volatile_acidity}{A numeric variable, ...}
     \item{citric_acid}{A numeric variable, ...}
     \item{residual_sugar}{A numeric variable, ...}
     \item{chlorides}{A numeric variable, ...}
     \item{free_sulphur_dioxide}{A numeric variable, ...}
     \item{total_sulphur_dioxide}{A numeric variable, ...}
     \item{density}{A numeric variable, ...}
     \item{pH}{A numeric variable, ...}
     \item{sulphates}{A numeric variable, ...}
     \item{alcohol}{A numeric variable, ...}
     \item{quality}{A numeric variable, ...}
}
}
\usage{
vinho_verde
}
\description{
White wine quality data related to variants of the Portuguese
     "Vinho Verde" wine. For more details, consult:
     \url{http://www.vinhoverde.pt/en/} or the reference Cortez et al., 2009:
}
\details{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis. Modelling
     wine preferences by data mining from physicochemical properties.
     In Decision Support Systems, Elsevier, 47(4):547-553. ISSN:0167-9236.
}
\keyword{datasets}
