% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{whiteside}
\alias{whiteside}
\title{whiteside}
\format{
A data frame with 56 rows and 3 columns:
\describe{
    \item{Insul}{factor: ... }
    \item{Temp}{numeric: ... }
    \item{Gas}{numeric: ... }
}
}
\usage{
whiteside
}
\description{
whiteside
}
\keyword{datasets}
