% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{worldLLhigh}
\alias{worldLLhigh}
\title{worldLLhigh}
\format{
A data frame with 187101 rows and 4 columns:
\describe{
    \item{PID}{integer: ... }
    \item{POS}{integer: ... }
    \item{X}{numeric: ... }
    \item{Y}{numeric: ... }
}
}
\usage{
worldLLhigh
}
\description{
worldLLhigh
}
\keyword{datasets}
