% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{wtloss}
\alias{wtloss}
\title{wtloss}
\format{
A data frame with 52 rows and 2 columns:
\describe{
    \item{Days}{integer: ... }
    \item{Weight}{numeric: ... }
}
}
\usage{
wtloss
}
\description{
wtloss
}
\keyword{datasets}
