% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxData.R
\docType{data}
\name{BirthWt}
\alias{BirthWt}
\title{BirthWt}
\format{
A data frame with 189 rows and 9 columns:
\describe{
    \item{low}{factor: ... }
    \item{age}{integer: ... }
    \item{lwt}{integer: ... }
    \item{race}{factor: ... }
    \item{smoke}{logical: ... }
    \item{ptl}{logical: ... }
    \item{ht}{logical: ... }
    \item{ui}{logical: ... }
    \item{ftv}{factor: ... }
}
}
\usage{
BirthWt
}
\description{
BirthWt
}
\keyword{datasets}
